+++ kornienko.artem/T3/commands.hpp
@@ -8,7 +8,7 @@ namespace kornienko
 {
   double evenOrOdd(const kornienko::Polygon & polygon, bool isEven);
   double mean(const kornienko::Polygon & polygon);
-  double numOfVertexes(const kornienko::Polygon & polygon, int num);
+  double numOfVertexes(const kornienko::Polygon & polygon, size_t num);
   void area(std::istream &, std::ostream & out, const std::vector< Polygon > polygons);
 }
